Allow different arguments for multiple instances of a pass (#6687)
Each pass instance can now store an argument for it, which can be different.

This may be a breaking change for the corner case of running a pass multiple
times and setting the pass's argument multiple times as well (before, the last
pass argument affected them all; now, it affects the last instance only). This
only affects arguments with the name of a pass; others remain global, as
before (and multiple passes can read them, in fact). See the CHANGELOG for
details.

Fixes #6646