Make FunctionInfo's assignment operator argument const
Aside from the fact that there's no need for this to be non-const and this is
the usual way to write an assignment operator, this is also needed because of a
recent change to std::pair (https://github.com/llvm/llvm-project/pull/89652).
This seems to be forcing pair to want the const version of the assignment
operator of its members.