Complete the support for try_table.
This includes adding interpreter support for `exnref` values.
Since they must carry a `Name tag` in addition to `Literals payload`,
they did not fit into existing `GCData`. Therefore, we add another
possible type in the big `union` for `ExnData`.

Alternatively, we could make a non-`funcref` `Name` value a valid
choice for a single `Literal`, but that would require inventing a
fake `Type` for such literals, such as `Type::tag`, which would
have no spec equivalent.

This highlighted the fact that using `Name`s will not work when
cross module boundaries (for `funcref` nor `exnref` nor the thrown
`WasmException`s). Since fixing this is beyond the scope of this
commit, we leave TODO's in place.

Other than the above, the changes are straightforward. They are
inspired from existing handling of `Try` and `Break`.