[NFC] Avoid v128 in rec groups with no other v128 uses (#6843)
We don't properly validate that yet. E.g.:

(module
 (rec
  (type $func (func))
  (type $unused (sub (struct (field v128))))
 )
 (func $func (type $func))
)

That v128 is not used, but it ends up in the output because it is in a rec group that is used.
Atm we do not require that SIMD be enabled in such a case, which can trip up the fuzzer.

Context: #6820. For now, modify the test that uncovered this.