[NFC] Avoid quadratic time when precomputing blocks (#6862)
When precomputing fails on a child block of a parent block, there is no point
to precompute the parent, as that will fail as well.

This makes --precompute on Emscripten's test_biggerswitch go from 1.44
seconds to 0.02 seconds (not a typo, that is 72x faster). The absolute number
is not that big, but we do run this pass more than once, so it saves a noticeable
chunk of time.