[NFC] Optimize ParamUtils::getUsedParams() (#6866)
This constructed a LocalGraph, which computes the sets that reach each get. But
all we need to know is which params are live, so instead we can do a liveness
computation (which is just a boolean, not the list of sets). Also, it is simple to get
the liveness computation to only work on the parameters and not all the locals,
as a further optimization.

Existing tests cover this, though I did find that the case of unreachability needed
a new test.

On a large testcase I am looking at, this makes --dae 17% faster.