Replace the old topological sort everywhere
To avoid having two separate topological sort utilities in the code base,
replace remaining uses of the old DFS-based, CRTP topological sort with the
newer Kahn's algorithm implementation.

This would be NFC, except that the new topological sort produces a different
order than the old topological sort, so the output of some passes is reordered.