[NFC] Convert HeapStoreOptimization to use a CFG (#6896)
This does not use the CFG yet, so there is no benefit (and likely some small
slowdown). The next PR will actually use it to fix a correctness bug. This PR
only sets up the CFG and converts the pass to operate on it, without changing
any behavior or tests.

Followup to #6882