[EH] Rename BrTarget to Trampoline (NFC) (#6898)
This renames "delegate_br_target" to "delegate_trampoline". So how we
translate `try`-`delegate` is:
- Before:
```wast
(try $delegate_target
  ...
  (try
    (do
      ...
    )
    (delegate $delegate_target)
  )
  ...
)
```

- After:
```wast
(try_table $delegate_target
  (throw_ref
    (block $delegate_br_target
      ...
      (try_table (catch_all $delegate_br_target)
        ...
      )
      ...
    )
  )
)
```

So `delegate_br_target` is the destination we branch (via `try_table`)
to, in order to rethrow the exnref using `throw_ref`.

But given that the translated code does not actually have a `br`, I
think this name can be confusing.

This renames `br_target` to `trampoline`, given that the block is upon
which we bounce the exnref off to reach the real delegate target.

This is to be consistent with the variable names in the LLVM
implementation (which has not been submitted yet).