Avoid conflicts with public rec groups in MinimizeRecGroups
As with all type optimizations, MinimizeRecGroups only changes private types,
which are the only types that are safe to modify. However, it is important for
correctness that MinimimizeRecGroups maintain separate type identities for all
types, whether public or private, to ensure that casts that should differentiate
two types cannot change behavior.

Previously the pass worked exclusively on private types, so there was nothing
preventing it from constructing a minimial rec group that happened to have the
same shape, and therefore type identity, as a public rec group. #6886 exhibits a
fuzzer test case where this happens and changes the behavior of the program.

Fix the bug by recording all public rec group shapes and resolve conflicts with
these shapes by updating the shape of the conflicting non-public type.

Fixes #6886.