Adds a J2CL specific pass that moves itable entries to vtables (#6888)
This allows to remove a reference field from all Java objects reducing
the per object memory and initialization overhead.

The pass is designed to run direclty on the J2CL output before other
optimizations since it relies on invariants that might get lost in
optimization. If the invariants don't hold the pass aborts.