[NFC-ish] Remove LocalGraph from LocalSubtyping (#6921)
The LocalGraph there was used for two purposes:

1. Get the list of gets and sets.
2. Get only the reachable gets and sets.

It is trivial to get all the gets and sets in a much faster way, by just walking the
code as this PR does. The downside is that we also consider unreachable gets
and sets, so unreachable code can prevent us from optimizing, but that seems
worthwhile as many passes make that assumption (and they all become
maximally effective after --dce). That is the only non-NFC part here.

Removing LocalGraph + the fixup code for unreachability makes this
significantly shorter, and also 2-3x faster.