[NFC] OptimizeAddedConstants: Early exit if there are no memories (#6926)
The pass optimizes loads and stores, so without a memory there is nothing to
do.

This only helps if the user set --low-memory-unused and also has no memory,
which is likely rare, but it's a trivial change so it seems worthwhile. In particular
this pass constructs a LocalGraph, so if we can avoid work it can be substantial.