[NFC] Optimize propagateLocals() (#6928)
That is the slowest part of --precompute-propagate, which is one of our slowest
passes. This makes that pass 25% faster.

The main change is to make the maps consider missing elements as non-constant,
rather than storing a None element in them. That saves allocating entries for the
common case of a non-constant set/get.

Also switch to a simple vector for the work queue, which is possible if we only
add work when a get/set becomes constant (which can only happen once). Another
benefit to adding work in that manner is that we don't start by adding every single
get/set as "work" at the start.