[wasm-split] Simplify handling of --keep-funcs and --split-funcs (#6948)
Maintain the invariant that every defined functions belongs to either
the set of kept functions or the set of split functions. Functions are
kept by default except when --keep-funcs is specified without
--split-funcs on the command line. This is mostly NFC except that it
changes the default behavior when no arguments are specified on the
command line to keep all functions.

This will simplify a follow-on PR that switches from passing the kept
functions to the module splitting utility to passing the split
functions.